+++ HISTORY.md
@@ -2,11 +2,12 @@
 
 ## in development:
 * **Notable Fixes:**
-    * fixed 'distribution' list selector support for nested atomic lists
+    * fixed 'distribution' and 'unique' list selectors support for nested atomic lists
 
 2023-11-21
 * new dev version for v9.0.14
 * threads: fixed 'distribution' list selector support for nested atomic lists
+* threads: fixed 'uniques' list selector support for nested atomic lists
 
 ## 9.0.13:
 * **Notable Fixes:**