+++ HISTORY.md
@@ -3,11 +3,14 @@
 ## in development:
 * **Notable Fixes:**
     * fixed 'distribution' and 'unique' list selectors support for nested atomic lists
+* **Translation Updates:**
+    * Spanish, thanks, David Martín!
 
 2023-11-21
 * new dev version for v9.0.14
 * threads: fixed 'distribution' list selector support for nested atomic lists
 * threads: fixed 'uniques' list selector support for nested atomic lists
+* Spanish translation update, thanks, David Martín!
 
 ## 9.0.13:
 * **Notable Fixes:**