+++ HISTORY.md
@@ -1,8 +1,10 @@
 # Snap! (BYOB) History
 
 ## in development:
+
+## 9.0.14:
 * **Notable Fixes:**
-    * fixed 'distribution' and 'unique' list selectors support for nested atomic lists
+    * fixed 'distribution' and 'uniques' list selectors support for nested atomic lists
 * **Translation Updates:**
     * Spanish, thanks, David Martín!
     * Chinese, thanks, moodykeke!
@@ -13,6 +15,7 @@
 * threads: fixed 'uniques' list selector support for nested atomic lists
 * Spanish translation update, thanks, David Martín!
 * Chinese translation update, thanks, moodykeke!
+* prepared v9.0.14 patch
 
 ## 9.0.13:
 * **Notable Fixes:**